pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexteco"
version = "0.1.0"
description = "Repository-native cost-of-running tooling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
nexteco = "nexteco.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["nexteco*"]
